Change u32 -> f32
Volume is a f32 value. (SwIPC describes it as a u32, but it is actually f32 as corroborated by switchbrew docs and SetAudioDeviceOutputVolume)

 ```cpp
const f32 volume = rp.Pop<f32>();
```

+++ src/core/hle/service/audio/audren_u.cpp
@@ -256,7 +256,7 @@ private:
 
         IPC::ResponseBuilder rb{ctx, 3};
         rb.Push(RESULT_SUCCESS);
-        rb.Push<u32>(1);
+        rb.Push<f32>(1);
     }
 
     void GetActiveAudioDeviceName(Kernel::HLERequestContext& ctx) {
